Organize stories in react example
