+++ examples/react/README.md
@@ -1,0 +1,5 @@
+# React
+
+This example demonstrates storybook in a React v16 project.
+
+It uses autotitle for some stories, but without a configuration object, a `titlePrefix` cannot be defined.
