+++ examples/react/stories/EnvironmentVariables.stories.jsx
@@ -1,6 +1,7 @@
 import { EnvironmentVariables } from './EnvironmentVariables';
 
 export default {
+  // This component uses auto-title
   component: EnvironmentVariables,
 };
