+++ examples/react/stories/components/Page.jsx
